+++ toutiao-service/src/main/java/com/tiaotiaopoker/service/WithdrawLogService.java
@@ -1,9 +1,11 @@
 package com.tiaotiaopoker.service;
 
+import java.math.BigDecimal;
+
 /**
  * Created by xiekang on 2018/9/24.
  */
 public interface WithdrawLogService {
 
-    int getAvailableWithdraw(String userId);
+    BigDecimal getAvailableWithdraw(String userId);
 }
